Inital refactor of db backend
