+++ tododb/database.go
@@ -1,0 +1,9 @@
+package tododb
+
+type TodoDB interface {
+	GetAllTodos() ([]string, error)
+	SaveTodo(string) error
+	DeleteTodo(string) error
+	GetHealthStatus() map[string]string
+	RegisterMetrics()
+}
